requirements.txt
networkx>=0.1
nose>=0.1
sniffer>=0.1
sympy>=0.1
python-intervals>=0.1